+++ derive_builder/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.11.1] - 2022-03-16
 - Forward `allow` and `cfg` attributes from the deriving struct to the builder and its impl block #222
 - Support passing attributes to the builder struct using `#[builder_struct_attr(...)]`
 - Support passing attributes to the builder struct's inherent `impl` block using `#[builder_impl_attr(...)]`